Implement companion assets
