+++ packages/remote-replay-launcher/src/index.ts
@@ -1,3 +1,4 @@
 export { executeRemoteTestRun } from "./execute-remote-test-run";
 export { uploadAssetsAndTriggerTestRun } from "./upload-assets-and-trigger-test-run";
+export { uploadAssets } from "./asset-upload-utils";
 export { TunnelData } from "./types";
